moved `HomeScreen` widget to separate file
